geweke convergence diagnostic

+++ corisummarizebigrun.py
@@ -3225,7 +3225,7 @@ def plotstarlc(flux,fluxerr,zpt,ids,mjd):
             plt.clf()
 
         ww = ids == id
-        axs.ravel()[int(i%100)].errorbar(np.float(mjd[ww]),flux[ww]*10**(.4*(31-zpt[ww])),yerr=fluxerr[ww]*10**(.4*(31-zpt[ww])),fmt='o',color='black')
+        axs.ravel()[int(i%100)].errorbar(mjd[ww],flux[ww]*10**(.4*(31-zpt[ww])),yerr=fluxerr[ww]*10**(.4*(31-zpt[ww])),fmt='o',color='black')
     pdf_pages.close()
     print 'saved allstarlc.pdf'
